Added details on requests page
The user can view thier own accounts requests and delete any if they choose too.

The home page requests that show up for each user doesn't include the logged in user anymore. When the offer button is clicked it now deducts from the balance of the user who created the request and adds it to the current logged in user.

gitignore file updated to ignore www/ folder