Names displayed instead of keys
The users name is shown in the drop down on the home page instead of a key and the myAccount page now gives the option to update user name and surname.

When an account is first made it will navigate to the myAccount page first instead of the home page.